foi feita a funcao que retorna as estatisticas das transacoes do ultimo minuto

+++ src/main/java/com/anjox/transacao/api/dto/request/RequestTransacaoDto.java
@@ -3,5 +3,5 @@
 import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
-public record RequestTransacaoDto(BigDecimal valor, OffsetDateTime dataHora) {
+public record RequestTransacaoDto(Double valor, OffsetDateTime dataHora) {
 }
